requirements.txt
feedparser>=6.0
google-genai>=1.0
supabase>=2.0
python-dotenv>=1.0
pyyaml>=6.0
